feat: approval app logout 刪除cache data,approval app 加入自動登入,socket 發射優化

+++ cryptoVerOne/LoginSignup/AuditViewController/subVC/SubListViewcontroller.swift
@@ -211,7 +211,7 @@ extension SubListViewcontroller:UITableViewDelegate,UITableViewDataSource
                 let y:CGFloat = offset.y + bouns.size.height - inset.bottom
                 let h:CGFloat = size.height
                 let reloadDistence = 50.0
-                let newH = size.height + reloadDistence + Views.tabBarHeight
+                let newH = size.height + CGFloat(reloadDistence) + Views.tabBarHeight
                 if y > newH {
                     tableView.tableFooterView = bottomRefrash
                     currentPage += 1
